requirements.txt
numpy
scipy
mpi4py
numba
python-dotenv